refactor(config): split crate modules
